+++ tests/NXPorts.Tests/PInvokeAssertions.cs
@@ -1,9 +1,5 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
-using System.Linq.Expressions;
 using System.Runtime.InteropServices;
 
 namespace NXPorts.Tests