+++ tests/NXPorts.Tests/TestFiles/Simple.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using NXPorts.Attributes;
+
+namespace NXPorts.Tests.TestFiles
+{
+    public static class Simple
+    {
+        [Export]
+        public static void DoSomething()
+        {
+            Console.WriteLine("Test");
+        }
+    }
+}